+++ analytics/models.py
@@ -1,20 +1,29 @@
 from django.db import models
 from django.urls import reverse
 from django.utils import timezone
-from django.conf import settings
 from django.contrib.auth.models import User
 from tinymce.models import HTMLField
+from django.core.mail import send_mail
+from khbr.settings import EMAIL_HOST_USER
+from django.template.loader import render_to_string
+from pathlib import Path
+
 class Author(models.Model):
-    user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, verbose_name='Пользователь')
-    bio = models.CharField('Немного о себе', max_length=500)
+    user = models.ForeignKey(User, on_delete=models.CASCADE, verbose_name='Пользователь')
+    slug = models.SlugField('Ссылка')
+    bio = HTMLField('Немного о себе')
+    avatar = models.ImageField(verbose_name='Аватар', upload_to='avatars/', blank=True)
 
     def __str__(self):
         return self.user.username
 
+    def get_absolute_url(self):
+        return reverse('authors_detail', kwargs={'slug': self.slug})
     
     class Meta:
         verbose_name = 'Автор'
         verbose_name_plural = 'Авторы'
+
 
 class Category(models.Model):
     name = models.CharField('Название', max_length=100)  
@@ -36,20 +45,44 @@ class Article(models.Model):
     title = models.CharField('Заголовок', max_length=120)
     slug = models.SlugField('Ссылка')
     short_description = HTMLField('Короткое описание на 200 символов', max_length=200)
-    image = models.ImageField("Фотография", upload_to='article/%Y/%m/%h/', blank=True)
+    image = models.ImageField("Фотография", upload_to='article/%Y/%m/%h/')
     content = HTMLField("Контент")
     integer_of_day = models.BooleanField('Цифра дня', default=False)
     author = models.ForeignKey(Author, on_delete=models.DO_NOTHING, verbose_name="Автор")
     created = models.DateTimeField('Дата создания', default=timezone.now)
     is_active = models.BooleanField(default=True, verbose_name='Видимость для пользователя')
     views = models.PositiveIntegerField('Просмотров публикации', default=0)
+    send_email = models.BooleanField(default=True, verbose_name='Отправить на рассылку')
 
     def __str__(self):
         return self.title
             
     def get_absolute_url(self):
-        return reverse('article_detail', kwargs={'slug': self.slug})
+        return reverse('article_detail', kwargs={'category_slug': self.category.slug, 'slug': self.slug})
     
+    def save(self):
+        if self.created:
+            subscribers = Subscribe.objects.values_list('email', flat=True)
+            if subscribers and self.send_email == True:
+                image = Path(self.image.url).name
+                html_message = render_to_string('../templates/mails/analytic.html', {'title': self.title,
+                                                                                    'short_description': self.short_description, 
+                                                                                    'category': self.category.name,
+                                                                                    'slug': self.slug,
+                                                                                    'category_slug': self.category.slug,
+                                                                                    'image': image})
+                send_mail(
+                        'Рассылка от KHBR - Аналитика',
+                        '',
+                        EMAIL_HOST_USER,
+                        subscribers,
+                        fail_silently=False,
+                        html_message=html_message
+                        )
+
+            super(Article, self).save()
+
+
     class Meta:
         verbose_name = 'Пост'
         verbose_name_plural = 'Посты'
@@ -75,3 +108,20 @@ class Subscribe(models.Model):
     class Meta:
         verbose_name = 'Подписчик на рассылку'
         verbose_name_plural = 'Подписчики на рассылку'
+
+class ArticleImage(models.Model):
+    article = models.ForeignKey(Article, on_delete=models.CASCADE, verbose_name='Пост')
+    title = models.CharField('Название картинки', blank=True, null=True, max_length=50)
+    image = models.ImageField(upload_to='article_inside/%Y/%m/%h/', verbose_name='Загрузить картинку')
+    author = models.ForeignKey(Author, on_delete=models.CASCADE, verbose_name="Автор")
+    created = models.DateTimeField('Дата создания', default=timezone.now)
+
+    def __str__(self):
+        if self.title:
+            return self.title
+        else:
+            return self.author.user.username
+
+    class Meta:
+        verbose_name = 'Фотография поста'
+        verbose_name_plural = 'Фотографии постов'