+++ users/forms.py
@@ -1,4 +1,5 @@
 from django import forms
+from django.forms import ModelForm
 from django.contrib.auth.models import User
 from django.contrib.auth.forms import UserCreationForm
 
@@ -7,9 +8,11 @@ class SignUpUserForm(UserCreationForm):
     class Meta:
         model = User
         fields = ['username', 'email',
-                'password']
+                'password1', 'password2']
 
-class SignInUserForm(UserCreationForm):
+
+"""class SignInUserForm(UserCreationForm):
     class Meta:
         model = User
         fields = ['username', 'password']
+"""