+++ CHANGELOG.md
@@ -1,3 +1,87 @@
+## 1.1.2 [22.06.2020]
+## 23 branch
+### Added
+- Navigation links on detail page
+- Mail's templates
+- Slider on detail pages 
+- Errors pages
+
+### Refactored
+- Hide anim on banner at main index
+
+
+## 1.1.1 [20.06.2020]
+## 21 branch
+### Added
+- User cabinet 
+- Documentation 
+- Featured posts at user cabinet 
+
+## 1.1.0 [20.06.2020]
+## 20 branch
+### Added
+- Views URL's of signIn and signUp templates
+- Checkbox for subscribe while signUp
+- Check for empty request in search
+- Check for data while sighIn and signUp
+- Hide/show password on signIn form
+- Logout view
+
+### Refactored
+- Templates 
+- Null true for event date start && end time
+- Error box on signIn or signUp
+
+## 1.0.0 [19.06.2020]
+## 19 branch
+### Added
+- Views URL's for author detail and index
+
+### Changed 
+- Styles of some links at detail and index
+
+### Refactored
+- Nothing founded form at search index
+
+## 0.9.0 [18.06.2020]
+## 18 branch
+### Added
+- Search block at header
+
+### Refactored
+- Markup of header
+
+## 0.8.0 [18.06.2020]
+## 16 branch
+### Added
+- Views of article_by_category, URL's
+- Context_processors with category on header
+- Format of date on all templates
+- Check if data exist of data template(changed template to if data exist)
+- URL's on index to main block
+- Category on article index
+- Short description if not image exist in article index 
+
+### Changed
+- Templates
+- Method: get_absolute_url of category model 
+
+### Refactored
+- Fixed position on title at rating index
+
+## 0.7.0 [16.06.2020] 
+## 13-14 branches
+### Added 
+- Search views, URL
+- Search result title URLs
+
+### Changed
+- Format date of results
+- Search not found template
+
+### Deleted
+- Protocols of links in footer
+
 ## 0.6.0 [15.06.2020]
 ### Added 
 - Category index page
@@ -6,7 +90,6 @@
 ### Refactored
 - Ecranization
 - Title at events and ratings detail
-
 
 ## 0.5.0 [14.06.2020]
 ### Added
@@ -31,4 +114,3 @@
 ### Added
 - Feature poll template index and poll detail
 - Slider with poll on main index page
-